@@ -16,55 +16,57 @@
     "prepublish": "yarn run snyk-protect"
   },
   "devDependencies": {
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@inertiajs/inertia": "^0.1.0",
+    "@inertiajs/inertia-vue": "^0.1.0",
+    "axios": "^0.18",
     "cross-env": "^6.0.3",
     "cypress": "3.5.0",
-    "eslint": "^6.0.1",
-    "eslint-config-standard": "^14.0.0",
-    "eslint-plugin-import": ">=2.13.0",
-    "eslint-plugin-node": ">=7.0.0",
-    "eslint-plugin-promise": ">=4.0.0",
-    "eslint-plugin-standard": ">=4.0.0",
-    "eslint-plugin-vue": "^6.0.0",
-    "mocha": "^6.0.2",
-    "mocha-junit-reporter": "^1.17.0",
-    "mocha-multi-reporters": "^1.1.7",
-    "moment-locales-webpack-plugin": "^1.0.7",
+    "eslint": "^6.1.0",
+    "eslint-plugin-vue": "^5.2.2",
+    "husky": "^2.7.0",
+    "laravel-mix": "^4.0.7",
+    "laravel-mix-purgecss": "^4.1.0",
+    "lint-staged": "^8.2.1",
+    "lodash": "^4.17.13",
+    "mrm": "^1.2.2",
+    "popper.js": "^1.12",
+    "postcss-import": "^12.0.1",
+    "postcss-nesting": "^7.0.0",
+    "prettier": "^1.18.2",
     "resolve-url-loader": "^3.1.1",
+    "sass": "^1.15.2",
     "sass-loader": "^8.0.0",
+    "tailwindcss": "^1.1.4",
+    "vue": "^2.5.21",
+    "vue-i18n": "^8.8",
     "vue-template-compiler": "^2.5.21"
   },
   "dependencies": {
-    "@hokify/vuejs-datepicker": "^2.0.2",
-    "animate.css": "^3.6.1",
-    "axios": "^0.19.0",
-    "bootstrap": "^4.0.0",
-    "font-awesome": "^4.7.0",
-    "hint.css": "^2.3.2",
-    "jquery": "^3.4.0",
-    "laravel-mix": "^5.0.0",
-    "laravel-mix-purgecss": "^4.2.0",
-    "list.js": "^1.5.0",
-    "lodash": "^4.17.14",
     "marked": "^0.8.0",
-    "moment": "^2.22.1",
-    "moment-timezone": "^0.5.21",
-    "popper.js": "^1.15.0",
-    "pretty-checkbox-vue": "^1.1.8",
-    "sass": "^1.15.2",
     "snyk": "^1.258.2",
-    "sweet-modal-vue": "^2.0.0",
-    "tachyons": "^4.9.1",
-    "v-selectmenu": "^2.0.1",
-    "vue": "^2.5.21",
-    "vue-autosuggest": "^2.0.0",
-    "vue-checkbox-radio": "^0.6.0",
-    "vue-clipboard2": "^0.3.0",
-    "vue-directive-tooltip": "^1.4.2",
-    "vue-good-table": "^2.13.0",
-    "vue-i18n": "^8.8",
-    "vue-js-toggle-button": "^1.2.2",
-    "vue-notification": "^1.3.6",
-    "vue-select": "^3.0.2"
+    "vue-meta": "^2.2.2"
+  },
+  "lint-staged": {
+    "linters": {
+      "resources/**/*.{css, scss}": [
+        "prettier --write",
+        "git add"
+      ],
+      "*.php": [
+        "php ./vendor/bin/php-cs-fixer fix --config .php_cs",
+        "git add"
+      ],
+      "*.vue": [
+        "eslint --ext .js,.vue resources/js/ --fix",
+        "git add"
+      ]
+    }
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
   },
   "snyk": true
 }
