@@ -1,34 +1,39 @@
-let mix = require('laravel-mix');
-require('laravel-mix-purgecss');
+const cssImport = require('postcss-import')
+const cssNesting = require('postcss-nesting')
+const mix = require('laravel-mix')
+const path = require('path')
+const tailwindcss = require('tailwindcss')
+const purgeCss = require('laravel-mix-purgecss');
 
-const MomentLocalesPlugin = require('moment-locales-webpack-plugin');
-mix.webpackConfig({
-    plugins: [
-        new MomentLocalesPlugin({
-            localesToKeep: ['en', 'ar', 'cs', 'de', 'es', 'fr', 'he', 'hr', 'it', 'nl', 'pt', 'pt-BR', 'ru', 'tr', 'zh-cn'],
-        }),
-    ],
-});
-
-let purgeCssOptions = {
-    whitelistPatterns: [/^fa-/, /^vdp-datepicker/, /^StripeElement/, /^vgt/, /^vue-tooltip/, /^pretty/],
-    whitelistPatternsChildren: [/^vdp-datepicker/, /^vgt/, /^vue-tooltip/, /^pretty/]
-};
+/*
+ |--------------------------------------------------------------------------
+ | Mix Asset Management
+ |--------------------------------------------------------------------------
+ |
+ | Mix provides a clean, fluent API for defining some Webpack build steps
+ | for your Laravel application. By default, we are compiling the Sass
+ | file for the application as well as bundling up all the JS files.
+ |
+ */
 
 mix.js('resources/js/app.js', 'public/js')
-    .sass('resources/sass/app-ltr.scss', 'public/css')
-    .sass('resources/sass/app-rtl.scss', 'public/css')
-
-    // stripe
-    .js('resources/js/stripe.js', 'public/js')
-    .sass('resources/sass/stripe.scss', 'public/css')
-
-    // u2f
-    .scripts(['resources/js/vendor/u2f/u2f-api.js'], 'public/js/u2f-api.js')
-
-    // global commands
-    .purgeCss(purgeCssOptions)
-    .extract()
-    .setResourceRoot('../')
-    .sourceMaps(false)
-    .version();
+  .postCss('resources/sass/app.scss', 'public/css', [
+    cssImport(),
+    cssNesting(),
+    tailwindcss('tailwind.js'),
+  ])
+  .purgeCss()
+  .webpackConfig({
+    output: { chunkFilename: 'js/[name].js?id=[chunkhash]' },
+    resolve: {
+      alias: {
+        vue$: 'vue/dist/vue.runtime.esm.js',
+        '@': path.resolve('resources/js'),
+      },
+    },
+  })
+  .babelConfig({
+    plugins: ['@babel/plugin-syntax-dynamic-import'],
+  })
+  .version()
+  .sourceMaps();
